backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
openai
faiss-cpu
pypdf
python-dotenv
pyjwt 